fixed order processing check

+++ internal/accrual/utils/types/types.go
@@ -15,8 +15,8 @@ type (
 	}
 
 	ordersGoods struct {
-		Description string `json:"description"`
-		Price       int    `json:"price"`
+		Description string  `json:"description"`
+		Price       float32 `json:"price"`
 	}
 )
